pyproject.toml
[tool.poetry]
name = "jiractl"
version = "0.1.0"
description = ""
authors = ["SamoKopecky <sano.kopecky@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
jiractl = "jiractl.main:main"